requirement.txt
mysql.connector==9.1.0
flask==1.1.2
PrettyTable==0.2.13
python-dotenv==1.0.1